+++ src/sybil_extras/parsers/docutils_rst/group_all.py
@@ -4,6 +4,7 @@ This parser uses the docutils library.
 """
 
 from beartype import beartype
+from sybil.typing import Evaluator
 
 from sybil_extras.parsers.abstract.group_all import AbstractGroupAllParser
 
@@ -14,3 +15,21 @@ class GroupAllParser(AbstractGroupAllParser):
     A parser that groups all code blocks in a document without
     markup.
     """
+
+    def __init__(
+        self,
+        *,
+        evaluator: Evaluator,
+        pad_groups: bool,
+    ) -> None:
+        """
+        Args:
+            evaluator: The evaluator to use for evaluating the combined
+        region.
+            pad_groups: Whether to pad groups with empty lines.
+        """
+        super().__init__(
+            evaluator=evaluator,
+            pad_groups=pad_groups,
+            no_pad_separator_lines=2,
+        )